WIP: Path examples
